Changed jfloat to jdouble on one line in the pls.poin binder where it obviously did not follow the pattern.  This fixed the x01 single-precision problem with pls.poin, and I also tested that x01 still worked when double precision was configured for the back end.
svn path=/trunk/; revision=3275

+++ bindings/java/javabind.c
@@ -1226,7 +1226,7 @@ JNIEXPORT void JNICALL Java_plplot_core_PLStream_poin__I_3D_3DI(
     PLFLT *x, *y;
     int must_free_buffers = 0, i;
 
-    if (sizeof(PLFLT) == sizeof(jfloat)) {
+    if (sizeof(PLFLT) == sizeof(jdouble)) {
     /* Trick: The cast is here to shut up the compiler in the case where
      * PLFLT != float, in which case the /other/ branch is the one that is
      * actually executed. */
